add basic support for 3-legged OAuth
This change seems to work, but essentially makes each Connection work only for a single user. It would be better to support a way to use a single connection but allow different `:access_token` and `:access_token_secret` values for each call. I have an idea for an API design for that, but I'm not sure yet how best to implement it.

Another weakness of the code at this point relates to the fact that some context.io API calls *do not support* being used in 3-legged mode (which seems at least somewhat sensible, since not all API calls require interacting with the authorized service). Such services need to invoke `request` with a parameter that turns off 3-legged OAuth for that call ... but the only one I've identified and modified so far is EmailSettings.